fix(film): scanLimit formula, -999 sentinel, extract shared test builder
- scanLimit: used UNIVERSAL_MARKER_LEN (11) instead of MARKER_PREFIX_BITS (3),
  stopping the scan 8 bit-positions too early and silently dropping fire events
  near the end of each replication chunk
- lastBytePos sentinel: -999 replaced with Number.NEGATIVE_INFINITY to make
  intent explicit and avoid silent failure if DEDUP_PROXIMITY_BYTES grows large
- buildSingleFireEventBytes / buildFireEventData: identical helpers extracted to
  film-fire-event-builder.ts; both test files now import the shared version
